♻️ inch closer to a similar api to replit-py
